Add support for SCRAM-SHA-256 and SCRAM-SHA-512 (#588)
* Add support for SCRAM-SHA-256 and -512 in the keycloak authenticator
