+++ keycloak/sasl-plugin/src/main/resources/META-INF/services/org.keycloak.credential.hash.PasswordHashProviderFactory
@@ -16,3 +16,4 @@
 #
 io.enmasse.keycloak.spi.ScramSha1PasswordHashProviderFactory
 io.enmasse.keycloak.spi.ScramSha256PasswordHashProviderFactory
+io.enmasse.keycloak.spi.ScramSha512PasswordHashProviderFactory
